[SCRUM-59] refactor: [Refactor] Kakao 클라이언트에 타임아웃/에러 매핑 추가
