+++ src/main/java/com/sprint/otboo/weather/integration/kakao/KakaoApiProperties.java
@@ -1,4 +1,4 @@
-package com.sprint.otboo.common.config;
+package com.sprint.otboo.weather.integration.kakao;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
