+++ src/main/java/com/sprint/otboo/weather/integration/kakao/client/KakaoLocalClientImpl.java
@@ -1,5 +1,6 @@
 package com.sprint.otboo.weather.integration.kakao.client;
 
+import com.sprint.otboo.common.exception.weather.WeatherProviderException;
 import com.sprint.otboo.weather.integration.kakao.dto.KakaoCoord2RegioncodeResponse;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.http.HttpStatusCode;
@@ -22,15 +23,31 @@ public class KakaoLocalClientImpl implements KakaoLocalClient {
         String uri = UriComponentsBuilder.fromPath("/v2/local/geo/coord2regioncode.json")
             .queryParam("x", longitude)
             .queryParam("y", latitude)
-            .build()
             .toUriString();
 
         return kakao.get()
             .uri(uri)
             .retrieve()
-            .onStatus(HttpStatusCode::is4xxClientError, ClientResponse::createException)
-            .onStatus(HttpStatusCode::is5xxServerError, ClientResponse::createException)
+            .onStatus(HttpStatusCode::is4xxClientError, resp ->
+                resp.createException().flatMap(ex -> {
+                    int code = resp.statusCode().value();
+                    if (code == 429) { // Too Many Requests
+                        return reactor.core.publisher.Mono.error(WeatherProviderException.tooManyRequests());
+                    }
+                    return reactor.core.publisher.Mono.error(WeatherProviderException.badGateway());
+                })
+            )
+            .onStatus(HttpStatusCode::is5xxServerError, resp ->
+                resp.createException().flatMap(ex -> {
+                    int code = resp.statusCode().value();
+                    if (code == 504) { // Gateway Timeout
+                        return reactor.core.publisher.Mono.error(WeatherProviderException.timeout());
+                    }
+                    return reactor.core.publisher.Mono.error(WeatherProviderException.badGateway());
+                })
+            )
             .bodyToMono(KakaoCoord2RegioncodeResponse.class)
+            .timeout(java.time.Duration.ofSeconds(5)) // 호출 타임아웃
             .block();
     }
 }
