+++ src/main/resources/application.yaml
@@ -31,5 +31,5 @@ logging:
 
 kakao:
   api:
-    base-url: ${KAKAO_API_BASE_URL:https://dapi.kakao.com}
-    rest-api-key: ${KAKAO_REST_API_KEY:__dummy__}
+    base-url: ${KAKAO_API_BASE_URL}
+    rest-api-key: ${KAKAO_REST_API_KEY}
